Moved descriptor metadata into the json file itself
Descriptor metadata which includes the type, guid and part will now be
part of the json format with a dollar-sign prepended inside the root
object. The type is also no longer a hex encoded id and instead is now
represented by the qualified name of the file which can be found in the
reflection-types.json file.

Additionally, an alignment issue has been fixed in the container file
format. (there may still be some lingering around)