Fixed and added new stuff to the cli tool
# Parser
Fixed issue with subsequent repacks.
Optimized metadata queries for kfc files. (not final)

# CLI
Added the ability to use stdout for unpack and stdin for repack by using
`--stdout` and `--stdin` respectively.
Added proper detection for .kfc and .exe files which means that it
should now also work with server files. There is also a `file-name`
argument which can be used to manually specify the name that should be
used.
Added kfc backup file validation which recreates the backup file, should
the game version differ.